feat: Introduce new styling system for markdown elements
- Added MarkdownTableStyle for configuring markdown table properties.
- Introduced MarkdownTextStyle for styling markdown text with support for animations and variants.
- Created SlideStyle to manage overall slide styling, including headings, text, and complex markdown elements.
- Established a new styling structure in the `styling.dart` file to export all relevant specs and styles.
- Updated theme management in `theme.dart` for a consistent dark theme across the application.
- Refactored imports across the codebase to align with the new styling structure.
- Removed obsolete style config loader tests and added comprehensive tests for the new style configuration system.
