+++ packages/superdeck/lib/src/rendering/blocks/block_widget.dart
@@ -6,7 +6,7 @@ import 'package:superdeck/src/rendering/blocks/block_provider.dart';
 import 'package:superdeck_core/superdeck_core.dart';
 
 import '../../deck/slide_configuration.dart';
-import '../../styling/styles.dart';
+import '../../styling/styling.dart';
 import '../../ui/widgets/error_widgets.dart';
 import '../../ui/widgets/provider.dart';
 import '../../utils/converters.dart';
