+++ packages/superdeck/lib/src/rendering/slides/slide_view.dart
@@ -3,7 +3,7 @@ import 'package:mix/mix.dart';
 import 'package:superdeck_core/superdeck_core.dart';
 
 import '../../deck/slide_configuration.dart';
-import '../../styling/styles.dart';
+import '../../styling/styling.dart';
 import '../../utils/constants.dart';
 import '../blocks/block_widget.dart';
 import 'package:flutter/widgets.dart';
