+++ packages/superdeck/lib/src/styling/specs/markdown_table_spec.dart
