+++ packages/superdeck/lib/src/styling/specs/slide_spec.dart
