--- packages/superdeck/lib/src/styling/styles.dart
@@ -1,29 +1,0 @@
-// Mix 2.0 Hand-written Specs
-export 'markdown_alert_spec.dart';
-export 'markdown_alert_type_spec.dart';
-export 'markdown_blockquote_spec.dart';
-export 'markdown_checkbox_spec.dart';
-export 'markdown_codeblock_spec.dart';
-export 'markdown_list_spec.dart';
-export 'markdown_table_spec.dart';
-export 'markdown_text_spec.dart';
-export 'slide_spec.dart';
-
-// Default styles
-export 'style.dart';
-
-// Mix 2.0 Hand-written Styles
-export 'markdown_alert_style.dart';
-export 'markdown_alert_type_style.dart';
-export 'markdown_blockquote_style.dart';
-export 'markdown_checkbox_style.dart';
-export 'markdown_codeblock_style.dart';
-export 'markdown_list_style.dart';
-export 'markdown_table_style.dart';
-export 'markdown_text_style.dart';
-export 'slide_style.dart';
-
-// Style Schema (YAML configuration)
-export 'schema/style_config_loader.dart';
-export 'schema/style_options_merger.dart';
-export 'schema/style_schemas.dart';
