+++ packages/superdeck/lib/src/styling/styles/markdown_text_style.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mix/mix.dart';
 
-import 'markdown_text_spec.dart';
+import '../specs/markdown_text_spec.dart';
 
 /// Style class for configuring [MarkdownTextSpec] properties.
 ///
