+++ packages/superdeck/lib/src/styling/styles/slide_style.dart
@@ -2,13 +2,13 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mix/mix.dart';
 
-import 'markdown_alert_spec.dart';
-import 'markdown_blockquote_spec.dart';
-import 'markdown_checkbox_spec.dart';
-import 'markdown_codeblock_spec.dart';
-import 'markdown_list_spec.dart';
-import 'markdown_table_spec.dart';
-import 'slide_spec.dart';
+import '../specs/markdown_alert_spec.dart';
+import '../specs/markdown_blockquote_spec.dart';
+import '../specs/markdown_checkbox_spec.dart';
+import '../specs/markdown_codeblock_spec.dart';
+import '../specs/markdown_list_spec.dart';
+import '../specs/markdown_table_spec.dart';
+import '../specs/slide_spec.dart';
 import 'markdown_alert_style.dart';
 import 'markdown_blockquote_style.dart';
 import 'markdown_checkbox_style.dart';
