+++ packages/superdeck/lib/src/ui/widgets/hero_element.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:mix/mix.dart';
 
-import '../../styling/styles.dart';
+import '../../styling/styling.dart';
 
 /// Immutable data holder for text elements in Hero animations
 class TextElement {
