+++ packages/superdeck/test/test_helpers.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:superdeck/src/rendering/slides/slide_view.dart';
 import 'package:superdeck/src/ui/widgets/provider.dart';
-import 'package:superdeck/src/styling/styles.dart';
+import 'package:superdeck/src/styling/styling.dart';
 import 'package:superdeck/src/deck/slide_configuration.dart';
 import 'package:superdeck/src/deck/widget_definition.dart';
 import 'package:superdeck_core/superdeck_core.dart';
