+++ packages/superdeck/test/testing_utils.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:superdeck/src/styling/styles.dart';
+import 'package:superdeck/src/styling/styling.dart';
 import 'package:superdeck/src/deck/slide_configuration.dart';
 import 'package:superdeck_core/superdeck_core.dart';
 export 'fixtures/slide_fixtures.dart';
